+++ derive/src/lib.rs
@@ -9,11 +9,6 @@
 #![warn(clippy::all, clippy::pedantic)]
 #![allow(clippy::use_self)]
 
-#[cfg(all(feature = "try_trait", not(feature = "unstable")))]
-compile_error!(
-    "The `try_trait` feature requires the `unstable` feature as an explicit opt-in to unstable features"
-);
-
 #[cfg(all(feature = "futures", not(feature = "unstable")))]
 compile_error!(
     "The `futures` feature requires the `unstable` feature as an explicit opt-in to unstable features"
@@ -32,16 +27,6 @@ compile_error!(
 #[cfg(all(feature = "trusted_len", not(feature = "unstable")))]
 compile_error!(
     "The `trusted_len` feature requires the `unstable` feature as an explicit opt-in to unstable features"
-);
-
-#[cfg(all(feature = "exact_size_is_empty", not(feature = "unstable")))]
-compile_error!(
-    "The `exact_size_is_empty` feature requires the `unstable` feature as an explicit opt-in to unstable features"
-);
-
-#[cfg(all(feature = "read_initializer", not(feature = "unstable")))]
-compile_error!(
-    "The `read_initializer` feature requires the `unstable` feature as an explicit opt-in to unstable features"
 );
 
 extern crate proc_macro;