+++ tests/enum_derive.rs
@@ -1,9 +1,6 @@
-#![cfg_attr(feature = "try_trait", feature(try_trait))]
 #![cfg_attr(feature = "generator_trait", feature(generator_trait))]
 #![cfg_attr(feature = "fn_traits", feature(fn_traits, unboxed_closures))]
 #![cfg_attr(feature = "trusted_len", feature(trusted_len))]
-#![cfg_attr(feature = "exact_size_is_empty", feature(exact_size_is_empty))]
-#![cfg_attr(feature = "read_initializer", feature(read_initializer))]
 #![cfg_attr(not(feature = "std"), no_std)]
 #![warn(unsafe_code)]
 #![warn(rust_2018_idioms)]
@@ -136,36 +133,6 @@ fn fn_traits() {
 #[test]
 fn trusted_len() {
     #[enum_derive(TrustedLen)]
-    enum Enum1<A, B> {
-        A(A),
-        B(B),
-    }
-}
-
-#[cfg(feature = "try_trait")]
-#[test]
-fn try_trait() {
-    #[enum_derive(Iterator)]
-    enum Enum1<A, B> {
-        A(A),
-        B(B),
-    }
-}
-
-#[cfg(feature = "exact_size_is_empty")]
-#[test]
-fn exact_size_is_empty() {
-    #[enum_derive(ExactSizeIterator)]
-    enum Enum1<A, B> {
-        A(A),
-        B(B),
-    }
-}
-
-#[cfg(feature = "read_initializer")]
-#[test]
-fn read_initializer() {
-    #[enum_derive(Read)]
     enum Enum1<A, B> {
         A(A),
         B(B),